build: rebuild `package-lock.json`
The existing `package-lock.json` included `fsevents` which is a macOS specific package.
This was causing errors when running a clean install (`npm ci`) on Linux hosts.
By deleting the existing `package-lock.json` and rebuilding it using `npm install --legacy-peer-deps`
on a Linux host machine, the `fsevents` error is resolved. This has the side-effect of updating
all of the pinned package versions to the latest versions that satisfy our `package.json` requirements.